Refactor reports out of pages and into reports-specific files
