+++ wagtail/admin/templates/wagtailadmin/reports/locked_pages.html
@@ -7,6 +7,6 @@
     {% include "wagtailadmin/shared/header.html" with title=title icon="locked" %}
 
     <div id="page-results">
-        {% include "wagtailadmin/pages/locked_pages_results.html" %}
+        {% include "wagtailadmin/reports/locked_pages_results.html" %}
     </div>
 {% endblock %}
