+++ wagtail/admin/views/pages.py
@@ -1285,14 +1285,3 @@ def revisions_unschedule(request, page_id, revision_id):
         'next': next_url,
         'subtitle': subtitle
     })
-
-
-def locked_pages(request):
-    pages = UserPagePermissionsProxy(request.user).editable_pages().filter(locked=True)
-
-    paginator = Paginator(pages, per_page=10)
-    pages = paginator.get_page(request.GET.get('p'))
-
-    return render(request, 'wagtailadmin/pages/locked_pages.html', {
-        'pages': pages,
-    })
